refactor(research): extract EPUB integrity validation
